requirements.txt
streamlit
scikit-learn
matplotlib
pandas
numpy
azure-storage-blob
statsmodels
